+++ docs/changelog.rst
@@ -2,12 +2,20 @@
 Changelog
 =========
 
+2.0.3
+=====
+
+* Fix label adaptation. Fix #256
+* Fix wrong radar truncation. Fix #255
+
+
 2.0.2
 =====
 
 * Fix view box differently to avoid getting a null height on huge numbers. (#254)
 * Fix broken font_family default
 * Fix non namespaced svg (without embed) javascript by adding uuid in config object. (config is in window.pygal now).
+
 
 2.0.1
 =====